add some security checks when running SUID root
on Darwin, htop needs to run with root privileges to display information
about other users processes. This commit makes running htop SUID root a
bit more safe.

+++ Process.c
@@ -513,12 +513,16 @@ void Process_toggleTag(Process* this) {
 }
 
 bool Process_setPriority(Process* this, int priority) {
-   int old_prio = getpriority(PRIO_PROCESS, this->pid);
-   int err = setpriority(PRIO_PROCESS, this->pid, priority);
-   if (err == 0 && old_prio != getpriority(PRIO_PROCESS, this->pid)) {
-      this->nice = priority;
+   if ( Process_getuid == 0 || Process_getuid == (int) this->st_uid ) {
+      int old_prio = getpriority(PRIO_PROCESS, this->pid);
+      int err = setpriority(PRIO_PROCESS, this->pid, priority);
+      if (err == 0 && old_prio != getpriority(PRIO_PROCESS, this->pid)) {
+         this->nice = priority;
+      }
+      return (err == 0);
    }
-   return (err == 0);
+   else
+      return false;
 }
 
 bool Process_changePriorityBy(Process* this, size_t delta) {
@@ -526,7 +530,8 @@ bool Process_changePriorityBy(Process* this, size_t delta) {
 }
 
 void Process_sendSignal(Process* this, size_t sgn) {
-   kill(this->pid, (int) sgn);
+   if ( Process_getuid == 0 || Process_getuid == (int) this->st_uid )
+      kill(this->pid, (int) sgn);
 }
 
 long Process_pidCompare(const void* v1, const void* v2) {
